Improve fun parsing to allow the dropped () form
I couldn't figure out an easy way to parse the "dropped () form" of
infix fun cases mentioned in the Definition without using unbounded
lookahead, which was forbidden previously because of the old
architecture of the lexer and parser as well as me objecting to
unbounded lookahead on a personal level.

However, recent refactors have brought us Tokens that implement Copy and
Lexers that parse all of the bytes into Tokens upfront instead of
lazily. So, we can now have unbounded lookahead. So, use that to parse
the previously mentioned "dropped ()" form.

The error messages from this might be mildly atrocious. To try to make
things a little better, _first_ try to parse the rarely-used form,
_then_ try the more common form. Also, share as much as possible: e.g.
expression parsing for the body and checking for ": ty" and "=". Also,
the very code itself to parse "atpat vid atpat" can be and is shared
with the "wrapped () form".