Rework exhaustiveness checking
The original algorithm was fundamentally flawed, as seen in the simple
example

```
val _ =
  case (true, false) of
    (true, _) => 0
  | (_, false) => 1
```

which was erroneously determined to be an exhaustive match.

Elements from the original approach were adapted into this
implementation, which is a bit more complex. I opted to decrease the
laziness of generating the 'needs' a bit, which I thought made reasoning
about the implementation a bit easier (who knows).

My original plan for this rewrite was to fully remove laziness but this
is not possible because of types with infinitely many values.

So laziness is introduced when a datatype constructor has an argument,
since as I understand this is the only place infinite types can arise.

I consider it impossible to exhaust any of the special types int, word,
string and char in any way other than to have the last case be a
wildcard or variable.