Add rudimentary handling of type variables
Doesn't perform any closure operations, so this might be unsound. Or
maybe it just rejects valid programs. Also doesn't do the implicit
scoping of type variables at val (and fun) bindings.

In any case, there is more work to be done, but hopefully this works in
basic situations.

Note that the wonky `let mut cx_cl;` trick is to avoid a clone when
there are no new type variables to add to the `Cx`, and is inspired from
a /r/rust post I saw one time about implementing FizzBuzz in Rust.