Rm some type name weirdness
As the comment alludes to, I'm not entirely sure the way TyName and
TyNameSet as mentioned in the Definition interact with Millet's
implementation strategy of using `Sym`s.

Before this commit, the simple test

```sml
signature SIG = sig
  type t
end
structure S: SIG = struct
  type t = int
end
```

would fail. This is because the type name `t` would be entered into the
Basis's TyNameSet after the `signature SIG` declaration, and then in the
`structure S` declaration, the type name `t` would be not bound by the
signature as expected, but by the Basis.

I suspect the strategy of using `Sym`s for type names makes the whole
discussion in the Definition about renaming bound type names
unnecessary, since `Sym`s are already meant to be globally unique.

Further, I think that the discussion about adding the type names from
top-level declarations to the Basis might be implicitly handled by
carrying around a State which used to generated totally fresh,
never-before-used `Sym`s when necessary.

I think, though, that the Basis does still need to have a TyNameSet, so
that we know what type names are already bound by the Basis and which
are bound by an Env when we want to turn an Env into a Sig. I confess
I'm not totally sure, and my confidence that I understand totally what
the Definition is asking for has dropped a bit.