Avoid a bunch of unreachable
We should not add a new variant for what a meta var can be. That
required us to add a bunch of new logic/unreachable which were actually
reachable (who'd a thunk). Instead let's just put the extra bit of data
we need on the unsolved meta var data (optionally).